backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
hvac==0.11.2
pyjwt==2.4.0
requests==2.26.0
flask-cors==3.0.10
cryptography==42.0.8
